Migrated Module.test to TS
